requirements.txt
#insert python library 
chromadb
langchain
streamlit
python-dotenv
sentence-transformers
openai
langchain-community